Forms category completed
input-field and Text-area components have been added in the documentation.
