+++ components/forms/_input-fields/input-focus.hbs
@@ -1,0 +1,28 @@
+<form class="msds-forms">
+    <div class="container">
+        <div class="row">
+            <div class="col-12 col-md-6">
+                <div class="msds-input">
+                    <div class="msds-input__icon-wrapper">
+                        <input id="input-field-focus" type="text" class="msds-input__text-input" placeholder="Enter your name*" name="text-input-disabled-lg">
+                        <label for="text-input-disabled-lg" class="msds-input__label">Enter your name*</label>
+                    </div>
+                    <p class="msds-input__info-message">
+                        We'll never share your name to anyone.
+                    </p>
+                </div>
+            </div>
+            <div class="col-12 col-md-6">
+                <div class="msds-input msds-input--small">
+                    <div class="msds-input__icon-wrapper">
+                        <input id="input-field-focus" type="text" class="msds-input__text-input" placeholder="Enter your name*" name="text-input-disabled-sm">
+                        <label for="text-input-disabled-sm" class="msds-input__label">Enter your name*</label>
+                    </div>
+                    <p class="msds-input__info-message">
+                        We'll never share your name to anyone.
+                    </p>
+                </div>
+            </div>            
+        </div>
+    </div>
+</form>
