+++ components/forms/_input-fields/input-readonly.hbs
@@ -1,0 +1,28 @@
+<form class="msds-forms">
+    <div class="container">
+        <div class="row">
+            <div class="col-12 col-md-6">
+                <div class="msds-input">
+                    <div class="msds-input__icon-wrapper">
+                        <input type="text" readonly="readonly" class="msds-input__text-input" placeholder="Enter your e-mail*" value="Readonly !" name="text-input-readonly-lg">
+                        <label for="text-input-readonly-lg" class="msds-input__label">Enter your e-mail*</label>
+                    </div>
+                    <p class="msds-input__info-message">
+                        We'll never share your email with anyone else.
+                    </p>
+                </div>
+            </div>
+            <div class="col-12 col-md-6">
+                <div class="msds-input msds-input--small">
+                    <div class="msds-input__icon-wrapper">
+                        <input type="text" readonly="readonly" class="msds-input__text-input" placeholder="Enter your e-mail*" value="Readonly !" name="text-input-readonly-sm">
+                        <label for="text-input-readonly-sm" class="msds-input__label">Enter your e-mail*</label>
+                    </div>
+                    <p class="msds-input__info-message">
+                        We'll never share your email with anyone else.
+                    </p>
+                </div>
+            </div>
+        </div>
+    </div>
+</form>
