+++ components/forms/_input-fields/input-sizes.hbs
@@ -1,0 +1,30 @@
+<form class="msds-forms">
+    <div class="container">
+        <div class="row">            
+            <div class="col-12 col-md-6">
+                <h3>Large Size (default)</h3>
+                <div class="msds-input">
+                    <div class="msds-input__icon-wrapper">
+                        <input type="text" class="msds-input__text-input" placeholder="Enter your e-mail*" name="text-input-lg">
+                        <label for="text-input-lg" class="msds-input__label">Enter your e-mail*</label>
+                    </div>
+                    <p class="msds-input__info-message">
+                        We'll never share your email with anyone else.
+                    </p>
+                </div>
+            </div>
+            <div class="col-12 col-md-6">
+                <h3>Small Size</h3>
+                <div class="msds-input msds-input--small">
+                    <div class="msds-input__icon-wrapper">
+                        <input type="text" class="msds-input__text-input" placeholder="Enter your e-mail*" name="text-input-sm">
+                        <label for="text-input-sm" class="msds-input__label">Enter your e-mail*</label>
+                    </div>
+                    <p class="msds-input__info-message">
+                        We'll never share your email with anyone else.
+                    </p>
+                </div>
+            </div>            
+        </div>
+    </div>
+</form>
