+++ components/forms/_input-fields/input-state-success.hbs
@@ -1,0 +1,28 @@
+<form class="msds-forms">
+    <div class="container">
+        <div class="row">
+            <div class="col-12 col-md-6">
+                <div class="msds-input msds-input--success">
+                    <div class="msds-input__icon-wrapper">
+                        <input type="text" class="msds-input__text-input" placeholder="Enter your e-mail*" name="text-input-success-lg">
+                        <label for="text-input-success-lg" class="msds-input__label">Enter your e-mail*</label>
+                    </div>
+                    <p class="msds-input__info-message">
+                        Looks good!
+                    </p>
+                </div>
+            </div>
+            <div class="col-12 col-md-6">
+                <div class="msds-input msds-input--small msds-input--success">
+                    <div class="msds-input__icon-wrapper">
+                        <input type="text" class="msds-input__text-input" placeholder="Enter your e-mail*" name="text-input-success-sm">
+                        <label for="text-input-error-sm" class="msds-input__label">Enter your e-mail*</label>
+                    </div>
+                    <p class="msds-input__info-message">
+                        Looks good!
+                    </p>
+                </div>
+            </div>            
+        </div>
+    </div>
+</form>
