+++ components/forms/_radio-buttons/radio-button-global-definition.hbs
@@ -1,4 +1,4 @@
-<form class="msds-form">
+<form class="msds-forms">
     <div class="container">
         <div class="row">
             <div class="col-6">
