+++ components/forms/_select-input/select-input-required.hbs
@@ -1,4 +1,5 @@
-<div class="col-12 col-md-5">
+<div class="col-12 col-md-6">
+    <h3>Required</h3>
     <div id="input-required" class="msds-select-input">
         <select name="color" class="msds-select-input__select" aria-placeholder="test">
         </select>
