+++ components/forms/_select-input/select-input-small.hbs
@@ -1,11 +1,9 @@
-<div class="col-12 col-md-3 offset-md-1">
+<div class="col-12 col-md-6">
+    <h3>Small Size</h3>
     <div id="input-small" class="msds-select-input">
-        <select class="msds-select-input__select msds-select-input__select--small" aria-placeholder="test">
-        </select>
+        <select class="msds-select-input__select msds-select-input__select--small" aria-placeholder="test"></select>
         <div class="msds-select-input__icon">
-            <svg title="flat-icon">
-                <use href="{{ path '/msds-spritemap.svg' }}#right-arrow" />
-            </svg>
+            <svg title="flat-icon"><use href="{{ path '/msds-spritemap.svg' }}#right-arrow" /></svg>
         </div>
         <div class="msds-select-input__validation-msg msds-text-body-4 msds-text-danger-red-1"></div>
     </div>
