+++ components/forms/_text-areas/text-area-success.hbs
@@ -1,0 +1,32 @@
+<form class="msds-forms">
+    <div class="container">
+        <div class="row">
+            <div class="col-6">
+                <div class="msds-text-area">
+                    <textarea class="msds-text-area__text-input msds-text-area__text-input--success" placeholder="Enter your description*" maxlength="400" type="textarea"></textarea>
+                    <label class="msds-text-area__label" for="default-input">Enter your description*</label>
+                    <p class="msds-text-area__message">
+                        <span>Please be aware that the description should not be longer than 400 characters.</span> 
+                        <span class="word-counter"> 
+                            <span class="word-counter__current">0</span>
+                            <span> / 400</span>
+                        </span>
+                    </p>
+                </div>
+            </div>
+            <div class="col-6">
+                <div class="msds-text-area msds-text-area--small">
+                    <textarea class="msds-text-area__text-input msds-text-area__text-input--success" placeholder="Enter your description*" maxlength="400" type="textarea"></textarea>
+                    <label class="msds-text-area__label" for="default-input">Enter your description*</label>
+                    <p class="msds-text-area__message">
+                        <span>Please be aware that the description should not be longer than 400 characters.</span> 
+                        <span class="word-counter"> 
+                            <span class="word-counter__current">0</span>
+                            <span> / 400</span>
+                        </span>
+                    </p>
+                </div>
+            </div>            
+        </div>
+    </div>
+</form>
