+++ components/forms/_text-areas/text-area.hbs
@@ -1,4 +1,4 @@
-<form class="msds-form">
+<form class="msds-forms">
     <div class="container">
         <div class="row">
             <div class="col-5">
@@ -100,8 +100,7 @@
                         placeholder="{{placeholder}}" maxlength="400" type="textarea"></textarea>
                     <label class="msds-form__input-container-text-area-label"
                         for="default-input">{{placeholder}}</label>
-                    <p class="msds-form__input-container-text-area-message">{{message}} <span class="word-count"></span>
-                    </p>
+                    <p class="msds-form__input-container-text-area-message">{{message}} <span class="word-count"></span></p>
                 </div>
             </div>
 
@@ -112,8 +111,7 @@
                         placeholder="{{placeholder}}" maxlength="400" type="textarea"></textarea>
                     <label class="msds-form__input-container-text-area-label"
                         for="default-input">{{placeholder}}</label>
-                    <p class="msds-form__input-container-text-area-message">{{message}} <span class="word-count"></span>
-                    </p>
+                    <p class="msds-form__input-container-text-area-message">{{message}} <span class="word-count"></span></p>
                 </div>
             </div>
         </div>
