--- components/forms/text-areas/text-area.scss
@@ -1,177 +1,0 @@
-.msds-form {
-  &__input-container {
-    &-text-area {
-      font-family: 'Open Sans', sans-serif;
-      resize: none;
-      width: 100%;
-      height: 200px;
-      padding: 24px 20px;
-      font-size: 16px;
-      color: $gray-10;
-      border: 1px solid $gray-5;
-      border-radius: 2px;
-
-      &:hover {
-        background-color: $gray-2;
-      }
-
-      &:focus {
-        outline: 4px solid rgba(0, 153, 218, 0.5);
-        border: solid 1px $clear-blue;
-        padding: 24px 20px;
-      }
-
-      &:focus ~ label {
-        left: 21px;
-        top: 10px;
-        font-size: 10px;
-      }
-
-      &::placeholder {
-        color: transparent;
-      }
-
-      &:not(:placeholder-shown) ~ label {
-        left: 21px;
-        top: 10px;
-        font-size: 10px;
-      }
-
-      &--small {
-        height: 120px;
-        font-size: 12px;
-        line-height: 18px;
-        padding-left: 12px;
-        padding-right: 12px;
-
-        &:focus {
-          padding-left: 12px;
-        }
-        &:focus ~ label {
-          font-size: 9px;
-          left: 12px;
-        }
-
-        & ~ label {
-          left: 12px;
-          font-size: 12px;
-          top: calc(20% - 20px);
-        }
-
-        &:not(:placeholder-shown) ~ label {
-          left: 12px;
-          top: 10px;
-          font-size: 9px;
-        }
-      }
-
-      &--disabled {
-        &:hover {
-          background-color: $gray-3;
-        }
-
-        & ~ label {
-          color: $gray-7;
-        }
-        background: $gray-3;
-      }
-
-      &--readonly {
-        &:hover {
-          background-color: $gray-3;
-        }
-
-        & ~ label {
-          color: $black;
-        }
-        background: $gray-3;
-        pointer-events: none;
-      }
-
-      &--error {
-        border: 1px solid $danger-red-1;
-        & ~ p {
-          color: $danger-red-1;
-        }
-
-        & ~ p:before,
-        & ~ p:after {
-          content: '';
-          position: absolute;
-          right: 25px;
-          top: calc(50% - 17px);
-          height: 13px;
-          width: 2px;
-          background-color: $danger-red-1;
-        }
-
-        & ~ p:before {
-          transform: rotate(45deg);
-        }
-
-        & ~ p:after {
-          transform: rotate(-45deg);
-        }
-
-        &:focus {
-          outline: 4px solid rgba(255, 40, 90, 0.5);
-          border: 1px solid $danger-red-2;
-        }
-      }
-
-      &--success {
-        border: 1px solid $confirm-green;
-        & ~ p {
-          color: $confirm-green;
-        }
-
-        & ~ p:before,
-        & ~ p:after {
-          content: '';
-          position: absolute;
-          right: 15px;
-          width: 2px;
-          background-color: $confirm-green;
-        }
-
-        & ~ p:before {
-          top: calc(50% - 15px);
-          height: 13px;
-          right: 15px;
-          transform: rotate(45deg);
-        }
-
-        & ~ p:after {
-          top: calc(50% - 10px);
-          height: 7px;
-          right: 21px;
-          transform: rotate(-45deg);
-        }
-
-        &:focus {
-          outline: 4px solid rgba(60, 192, 3, 0.5);
-          border: 1px solid $confirm-green;
-        }
-      }
-    }
-  }
-  &__input-container-text-area-label {
-    position: absolute;
-    top: calc(15% - 23px);
-    left: 35px;
-    transition: 0.5s;
-    pointer-events: none;
-    color: $gray-8;
-    font-size: 16px;
-  }
-
-  &__input-container-text-area-message {
-    display: flex;
-    justify-content: space-between;
-    font-size: 12px;
-    line-height: 18px;
-    color: $gray-9;
-    margin-left: 3px;
-    margin-top: 6px;
-  }
-}
