--- docs/Content/Typography/-general-information.md
@@ -1,18 +1,0 @@
----
-title: General Information
----
-Typography is based on googles [open sans](https://fonts.google.com/specimen/Open+Sans)
-
-### Adding google font to page:
-Below is the link to include google fonts on the page
-```html
-<link href='//fonts.googleapis.com/css?family=Open+Sans:300italic,400italic,600italic,300,400,700,600' rel='stylesheet' type='text/css' />
-```
-
-### How to enable open sans font family on elements:
-Add 'msds-enable-font-family' class to the parent element e.g. below example of how to do it on the entire page
-```
-<html class="msds-enable-font-family">
-	......
-</html>
-```
