+++ docs/Content/Typography/Lists.md
@@ -4,12 +4,14 @@ title: Lists
 
 ## In this section, we are going to give you an overview of all the different web lists.
 
-Below are the typography definitions that you need to use for them on our CMS. Right now, only the unordered list has been defined and it should used as such.
+Below are the definitions that you need to use for all lists on our CMS. Right now, only the unordered list has been designed.
 
 ### Table of contents
 <div class="row">    
-    <div class="col-md-6">      
-        <li class="document__unordered-list-item"> <a class="msds-link" href="#lists">Lists</a></li>
+    <div class="col-md-6">  
+      <ul class="document__unordered-list">    
+        <li class="document__unordered-list-item"> <a class="msds-link" href="#overview">Overview</a></li>
+        <li class="document__unordered-list-item"> <a class="msds-link" href="#types">Types</a></li>
         <ul class="document__unordered-list">
           <li class="document__unordered-list-item">
             <a class="msds-link" href="#unordered-lists-definitions">Unordered Lists</a>
@@ -19,7 +21,14 @@ Below are the typography definitions that you need to use for them on our CMS. R
     </div>
 </div>
 
+## Overview 
+Our list definitions is only based on light themes and only one type of list has been defined at the moment.
+
+## Types 
+Different types for lists are available, however we have only designed the unordered list for light themes.
+
 ### Unordered Lists Definitions
+The unordered list should look like this on light themes. 
 
 <div class="element-preview">
   <div class="element-preview__inner">{{render '@typography-unordered-lists'}}</div>
