+++ src/js/design-system.js
@@ -4,6 +4,7 @@ import './main.js'
 //JS files for documnetations
 import 'components/_Progress-Bar/progress-bar-initialization.js'
 import 'components/forms/_select-input/select-input-initialization.js'
+import 'components/forms/_text-areas/text-area.js'
 import 'components/_expanding-components/accordion-initialization.js'
 import 'components/_expanding-components/section-initialization.js'
 import 'components/_icons/flat-icons.js'
