+++ src/scss/main.scss
@@ -15,8 +15,8 @@
 @import 'components/button-cards/_solution-card/solution-card.scss';
 @import 'components/button-cards/_basic-card/basic-card.scss';
 @import 'components/tab-tile/tab-tile.scss';
-@import 'components/forms/input-fields/input-field.scss';
-@import 'components/forms/text-areas/text-area.scss';
+@import 'components/forms/_input-fields/input-field.scss';
+@import 'components/forms/_text-areas/text-area.scss';
 @import 'components/forms/_checkboxes/checkbox.scss';
 @import 'components/forms/_radio-buttons/radio-button.scss';
 @import 'components/forms/_select-input/select-input.scss';
